Fix SparqlEngine build errors and restore zero-GC design
- Fix 56+ compilation errors across SparqlEngine project
- Add missing using statements (System.IO)
- Add constructors to readonly ref structs (ResolvedTriple, OptionalResult, PathResult, etc.)
- Fix ref struct scope issues with scoped parameters
- Add missing _tripleCount field to TemporalTripleStore
- Change private nested types to internal for accessibility
- Add QueryAsOf method to MultiTemporalStore

Zero-GC restorations:
- PropertyPath: Changed from class to ref struct with flat node array
- PropertyPathNode: Uses int offsets instead of string references
- ResourceDescriptor: Uses int offsets instead of string field
- Prologue.BaseUri: Uses int offsets instead of string field
- All parsing uses stackalloc and Span<T> for zero heap allocation

Also adds CLAUDE.md for Claude Code guidance.

+++ src/SparqlEngine/AtomStore.cs
@@ -17,7 +17,7 @@ public sealed unsafe class AtomStore : IDisposable
 {
     private const int PageSize = 4096; // 4KB pages
     private const int HashTableSize = 1 << 20; // 1M buckets
-    private const int InitialDataSize = 1L << 30; // 1GB initial
+    private const long InitialDataSize = 1L << 30; // 1GB initial
     
     // Memory-mapped files
     private readonly FileStream _dataFile;
